Even more memory access validation fixes
Even more memory access validation fixes! The correct memory access permissions should now be returned in every case.

In the event that an address range is entirly contained within a single region then the permissions for that region are returned.
In the event that an address range crosses two or more regions, the permissions of the region with the highest required permissions are returned. This should prove more security friendly, though in theory this shouldn't ever happen.