Modify the flags that are used by the multiply and divide instructions
The multiply instructions will now only change the overflow (OF) and carry (CF) flags. These will be set to 1 when the upper 16 bits are not zero. All other flags are undefined.

The divide instruction will now not change any flags, meaning they are all undefined. Also added a test to ensure that division by zero produces an exception.

Update the associated tests to reflect these changes.