Implement the basic boot ROM, unconditional jump instruction and associated changes.
Implement the basic boot ROM, which now handles the CPU setup. Added the unconditional jump instruction to make it feasible. On boot, the CPU will now look at memory address 0xDEADBEEF for the boot rom and will execute any code found there before jumping to the user compiled code. Delete some code that's no longer needed due to this change. There are more changes that will be added as more functionality is added - such enabling and disabling interrupts and adding an interrupt catcher.

Add some tests for the unconditional jump instruction.

Change the output of the print memory segments function to display the addresses in hex, for easier reading.