Add unconditional jump by register value instruction, update the boot ROM to use this function and swap the interrupt flag to be enabling interrupts rather than disabling them.
Add the unconditional jump instruction that uses a register value as a pointer. This was useful for the boot ROM.

Update the boot ROM to use the unconditional jump by register pointer instruction. We can now simply jump to the address as specifed by the code segment (CS) register.

Switch the interrupt CPU flag to be allowing interrupts when enabled as opposed to disabling them when enabled. This is more in line with how most CPUs handle it. Update the tests and boot ROM to reflect this change.