Rework instruction tests to avoid needing to check the entire user-space memory segment.
We don't want to do this because we need to ensure that the tests use a memory size that is at least the minumum permitted. This is because we always need to ensure that there is sufficient space for the interrupt vector, which will typically be found at memory address 0x0.

Changing this required an extensive rework of the tests, but it now means that we don't need to specify the memory contents as part of the test results.

This change also contains some initial work on the interrupt handling, but isn't fully hooked up yet.