A significant rework of the ASM parser.
The ASM parser will now select instructions by a type of fuzzy analysis with their argument types. If you have an instruction of the format: ins 0xaa, 0xaa and another ins 0xdeadbeef, 0xdeadbeef the parser will select the option that takes the least amount of space to store based on the arguments presented.

Where needed specific alternative names will be added to allow the parser to specifically target an instruction with a smaller instruction size.