Refactor the parser to better handle instructions that have labels.
The parser will handle instructions that have labels better now, and these instructions will use the label rather than the pseudo address when displayed.

In addition to this the code to calculate the destination addresses for local (relative) jumps has been removed, this will be added to the compiler instead.